+++ test/mod_test_fchange_normalization.f90
@@ -2,7 +2,7 @@
 !> @brief Unit test suite for calc_fchange routine.
 module mod_test_calc_fchange
   use asserts
-  use, intrinsic :: iso_fortran_env, only: real64
+  use, intrinsic :: iso_fortran_env, only: real64, int32
   implicit none
   public
 
@@ -39,7 +39,7 @@ contains
   !> @brief Run all calc_fchange tests.
   subroutine run_all_tests_calc_fchange()
     type(test_case) :: all_tests(10)
-    integer :: i
+    integer(int32) :: i
     
     all_tests = get_all_tests()
     
@@ -54,7 +54,7 @@ contains
   subroutine run_named_tests_calc_fchange(test_names)
     character(len=*), intent(in) :: test_names(:)
     type(test_case) :: all_tests(10)
-    integer :: i, j
+    integer(int32) :: i, j
     logical :: found
     
     all_tests = get_all_tests()
@@ -77,8 +77,8 @@ contains
 
   !> @brief Test basic fold change calculation (from R test).
   subroutine test_calc_fchange_basic_calculation()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(1) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(1) :: control_cols, cond_cols
     real(real64), dimension(4) :: i_matrix  ! 2 genes × 2 samples
     real(real64), dimension(2) :: o_matrix  ! 2 genes × 1 pair
     real(real64), dimension(2) :: expected_matrix
@@ -103,8 +103,8 @@ contains
 
   !> @brief Test result correctness with specific expected values (from R test).
   subroutine test_calc_fchange_result_correctness()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(1) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(1) :: control_cols, cond_cols
     real(real64), dimension(4) :: i_matrix
     real(real64), dimension(2) :: o_matrix
     
@@ -125,8 +125,8 @@ contains
 
   !> @brief Test that output dimensions are correct.
   subroutine test_calc_fchange_preserves_dimensions()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(2) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(2) :: control_cols, cond_cols
     real(real64), dimension(12) :: i_matrix  ! 3 genes × 4 samples
     real(real64), dimension(6) :: o_matrix   ! 3 genes × 2 pairs
     
@@ -148,8 +148,8 @@ contains
 
   !> @brief Test with single gene, single pair.
   subroutine test_calc_fchange_single_pair()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(1) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(1) :: control_cols, cond_cols
     real(real64), dimension(2) :: i_matrix  ! 1 gene × 2 samples
     real(real64), dimension(1) :: o_matrix  ! 1 gene × 1 pair
     
@@ -167,8 +167,8 @@ contains
 
   !> @brief Test with multiple condition-control pairs.
   subroutine test_calc_fchange_multiple_pairs()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(3) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(3) :: control_cols, cond_cols
     real(real64), dimension(12) :: i_matrix  ! 2 genes × 6 samples
     real(real64), dimension(6) :: o_matrix   ! 2 genes × 3 pairs
     
@@ -193,8 +193,8 @@ contains
 
   !> @brief Test with negative expression values.
   subroutine test_calc_fchange_negative_values()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(2) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(2) :: control_cols, cond_cols
     real(real64), dimension(8) :: i_matrix  ! 2 genes × 4 samples
     real(real64), dimension(4) :: o_matrix  ! 2 genes × 2 pairs
     
@@ -217,8 +217,8 @@ contains
 
   !> @brief Test with zero expression values.
   subroutine test_calc_fchange_zero_values()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(1) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(1) :: control_cols, cond_cols
     real(real64), dimension(4) :: i_matrix  ! 2 genes × 2 samples
     real(real64), dimension(2) :: o_matrix  ! 2 genes × 1 pair
     
@@ -238,8 +238,8 @@ contains
 
   !> @brief Test with large expression values for numerical stability.
   subroutine test_calc_fchange_large_values()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(1) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(1) :: control_cols, cond_cols
     real(real64), dimension(4) :: i_matrix  ! 2 genes × 2 samples
     real(real64), dimension(2) :: o_matrix  ! 2 genes × 1 pair
     
@@ -261,8 +261,8 @@ contains
 
   !> @brief Test with identical control and condition values (zero fold change).
   subroutine test_calc_fchange_identical_values()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(2) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(2) :: control_cols, cond_cols
     real(real64), dimension(8) :: i_matrix  ! 2 genes × 4 samples
     real(real64), dimension(4) :: o_matrix  ! 2 genes × 2 pairs
     
@@ -287,8 +287,8 @@ contains
 
   !> @brief Test with mixed positive, negative, and zero values.
   subroutine test_calc_fchange_mixed_values()
-    integer :: n_genes, n_cols, n_pairs
-    integer, dimension(3) :: control_cols, cond_cols
+    integer(int32) :: n_genes, n_cols, n_pairs
+    integer(int32), dimension(3) :: control_cols, cond_cols
     real(real64), dimension(18) :: i_matrix  ! 3 genes × 6 samples
     real(real64), dimension(9) :: o_matrix   ! 3 genes × 3 pairs
