+++ test/mod_test_log2_normalization.f90
@@ -2,7 +2,7 @@
 !> @brief Unit test suite for log2_transformation routine.
 module mod_test_log2_transformation
   use asserts
-  use, intrinsic :: iso_fortran_env, only: real64
+  use, intrinsic :: iso_fortran_env, only: real64, int32
   implicit none
   public
 
@@ -41,7 +41,7 @@ contains
   !> @brief Run all log2_transformation tests.
   subroutine run_all_tests_log2_transformation()
     type(test_case) :: all_tests(12)
-    integer :: i
+    integer(int32) :: i
     
     all_tests = get_all_tests()
     
@@ -56,7 +56,7 @@ contains
   subroutine run_named_tests_log2_transformation(test_names)
     character(len=*), intent(in) :: test_names(:)
     type(test_case) :: all_tests(12)
-    integer :: i, j
+    integer(int32) :: i, j
     logical :: found
     
     all_tests = get_all_tests()
@@ -79,7 +79,7 @@ contains
 
   !> @brief Test log2(x+1) transformation with basic known values (from R test).
   subroutine test_log2_basic_values()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat, expected_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
     
@@ -102,7 +102,7 @@ contains
 
   !> @brief Test that log2(0+1) = 0 for all zeros (from R test).
   subroutine test_log2_zeros_handling()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat, expected_flat
     
     n_genes = 2; n_tissues = 2
@@ -118,7 +118,7 @@ contains
 
   !> @brief Test that dimensions are preserved (from R test concept).
   subroutine test_log2_preserves_dimensions()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(6) :: input_flat, output_flat
     
     n_genes = 2; n_tissues = 3
@@ -135,7 +135,7 @@ contains
 
   !> @brief Test single element matrix.
   subroutine test_log2_single_element()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(1) :: input_flat, output_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
     
@@ -150,7 +150,7 @@ contains
 
   !> @brief Test with large values to check numerical stability.
   subroutine test_log2_large_values()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
     
@@ -170,7 +170,7 @@ contains
 
   !> @brief Test with very small positive values.
   subroutine test_log2_small_values()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
     
@@ -188,7 +188,7 @@ contains
 
   !> @brief Test with powers of 2 minus 1 for exact results.
   subroutine test_log2_powers_of_two()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat, expected_flat
     
     n_genes = 2; n_tissues = 2
@@ -204,11 +204,11 @@ contains
 
   !> @brief Test with random matrix for general properties.
   subroutine test_log2_random_matrix()
-    integer, parameter :: n_genes = 5, n_tissues = 4
+    integer(int32), parameter :: n_genes = 5, n_tissues = 4
     real(real64), dimension(n_genes * n_tissues) :: input_flat, output_flat
-    integer :: i
-    integer :: n_seed
-    integer, allocatable :: seed_array(:)
+    integer(int32) :: i
+    integer(int32) :: n_seed
+    integer(int32), allocatable :: seed_array(:)
     ! For reproducibility: initialize the random number generator seed
     call random_seed(size=n_seed)
     allocate(seed_array(n_seed))
@@ -233,7 +233,7 @@ contains
 
   !> @brief Test behavior with negative values (should still work due to +1).
   subroutine test_log2_negative_handling()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
     
@@ -251,7 +251,7 @@ contains
 
   !> @brief Test edge cases and boundary conditions.
   subroutine test_log2_edge_cases()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(4) :: input_flat, output_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
     
@@ -269,9 +269,9 @@ contains
 
   !> @brief Test monotonic property: log2(x+1) is strictly increasing.
   subroutine test_log2_monotonic_property()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(5) :: input_flat, output_flat
-    integer :: i
+    integer(int32) :: i
     
     n_genes = 5; n_tissues = 1
     input_flat = [1.0d0, 2.0d0, 5.0d0, 10.0d0, 20.0d0]  ! Increasing sequence
@@ -287,7 +287,7 @@ contains
 
   !> @brief Test mathematical properties of log2 transformation.
   subroutine test_log2_mathematical_properties()
-    integer :: n_genes, n_tissues
+    integer(int32) :: n_genes, n_tissues
     real(real64), dimension(8) :: input_flat, output_flat
     real(real64), parameter :: LOG2 = log(2.0d0)
